+++ apps/jaspr_pad/lib/providers/docu_provider.dart
@@ -1,4 +1,4 @@
-import 'package:jaspr_riverpod/jaspr_riverpod.dart';
+import 'package:jaspr_riverpod/legacy.dart';
 
 import '../models/api_models.dart';